+++ src/details/ArborX_DetailsUtils.hpp
@@ -13,6 +13,8 @@
 #define ARBORX_DETAILS_UTILS_HPP
 
 #include <ArborX_DetailsKokkosExtAccessibilityTraits.hpp>
+#include <ArborX_DetailsKokkosExtMinMaxOperations.hpp>
+#include <ArborX_DetailsKokkosExtStdAlgorithms.hpp>
 #include <ArborX_DetailsKokkosExtViewHelpers.hpp>
 #include <ArborX_Exception.hpp>
 
@@ -156,83 +158,20 @@ create_layout_right_mirror_view_and_copy(ExecutionSpace const &execution_space,
   }
 }
 
-// NOTE: This functor is used in exclusivePrefixSum( src, dst ).  We were
-// getting a compile error on CUDA when using a KOKKOS_LAMBDA.
-template <typename T, typename DeviceType>
-class ExclusiveScanFunctor
-{
-public:
-  ExclusiveScanFunctor(Kokkos::View<T *, DeviceType> const &in,
-                       Kokkos::View<T *, DeviceType> const &out)
-      : _in(in)
-      , _out(out)
-  {}
-  KOKKOS_INLINE_FUNCTION void operator()(int i, T &update,
-                                         bool final_pass) const
-  {
-    T const in_i = _in(i);
-    if (final_pass)
-      _out(i) = update;
-    update += in_i;
-  }
-
-private:
-  Kokkos::View<T *, DeviceType> _in;
-  Kokkos::View<T *, DeviceType> _out;
-};
 } // namespace Details
 
-/** \brief Computes an exclusive scan.
- *
- *  \param[in] space Execution space
- *  \param[in] src Input view with range of elements to sum
- *  \param[out] dst Output view; may be equal to \p src
- *
- *  When \p dst is not provided or if \p src and \p dst are the same view, the
- *  scan is performed in-place.  "Exclusive" means that the i-th input element
- *  is not included in the i-th sum.
- *
- *  \pre \p src and \p dst must be of rank 1 and have the same size.
- */
 template <typename ExecutionSpace, typename ST, typename... SP, typename DT,
           typename... DP>
-void exclusivePrefixSum(ExecutionSpace &&space,
-                        Kokkos::View<ST, SP...> const &src,
-                        Kokkos::View<DT, DP...> const &dst)
+[[deprecated]] void exclusivePrefixSum(ExecutionSpace &&space,
+                                       Kokkos::View<ST, SP...> const &src,
+                                       Kokkos::View<DT, DP...> const &dst)
 {
-  static_assert(
-      std::is_same<
-          typename Kokkos::ViewTraits<DT, DP...>::value_type,
-          typename Kokkos::ViewTraits<DT, DP...>::non_const_value_type>::value,
-      "exclusivePrefixSum requires non-const destination type");
-
-  static_assert(
-      (unsigned(Kokkos::ViewTraits<DT, DP...>::rank) ==
-       unsigned(Kokkos::ViewTraits<ST, SP...>::rank)) &&
-          (unsigned(Kokkos::ViewTraits<DT, DP...>::rank) == unsigned(1)),
-      "exclusivePrefixSum requires Views of rank 1");
-
-  using ValueType = typename Kokkos::ViewTraits<DT, DP...>::value_type;
-  using DeviceType = typename Kokkos::ViewTraits<DT, DP...>::device_type;
-
-  auto const n = src.extent(0);
-  ARBORX_ASSERT(n == dst.extent(0));
-  Kokkos::RangePolicy<std::decay_t<ExecutionSpace>> policy(
-      std::forward<ExecutionSpace>(space), 0, n);
-  Kokkos::parallel_scan(
-      "ArborX::Algorithms::exclusive_scan", policy,
-      Details::ExclusiveScanFunctor<ValueType, DeviceType>(src, dst));
+  Details::KokkosExt::exclusive_scan(std::forward<ExecutionSpace>(space), src,
+                                     dst);
 }
 
-/** \brief In-place exclusive scan.
- *
- *  \param[in] space Execution space
- *  \param[in,out] v View with range of elements to sum
- *
- *  Calls \c exclusivePrefixSum(v, v)
- */
 template <typename ExecutionSpace, typename T, typename... P>
-inline std::enable_if_t<
+[[deprecated]] inline std::enable_if_t<
     Kokkos::is_execution_space<std::remove_reference_t<ExecutionSpace>>::value>
 exclusivePrefixSum(ExecutionSpace &&space, Kokkos::View<T, P...> const &v)
 {
